requirements.txt
distlib==0.3.5
filelock==3.7.1
gunicorn==20.0.4
numpy==1.23.0
pandas==1.4.3
platformdirs==2.5.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
virtualenv==20.15.1
newsdataapi==0.1.5
yfinance==0.1.74
WTForms==3.0.1
form==0.0.1
Flask==2.1.3
Flask-WTF==1.0.1
itsdangerous==2.1.2
flask-behind-proxy==0.1.1
email-validator==1.2.1
Flask-Login==0.6.1